+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Bugfixes:
   - Suite Matrix does not work -> [View Issue](https://github.com/aquality-automation/aquality-tracking-ui/issues/18)
   - Local Manager can see Local Permissions page -> [View Issue](https://github.com/aquality-automation/aquality-tracking-ui/issues/22)
+  - Test View: Suite link leads to 404 page -> [View Issue](https://github.com/aquality-automation/aquality-tracking-ui/issues/39)
 
 ## 0.3.2 (2019-09-10)